requirements.txt
absl-py==0.9.0
aiofiles==0.5.0
aiohttp==3.6.2
alembic==1.4.2
aniso8601==8.0.0
APScheduler==3.6.3
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
boto3==1.14.11
botocore==1.17.11
cachetools==4.1.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
contextvars==2.4
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
fbmessenger==6.0.0
Flask==1.1.2
Flask-Cors==3.0.8
flask-restplus==0.13.0
future==0.18.2
gast==0.2.2
geocoder==1.38.1
geographiclib==1.50
geopy==2.0.0
gevent==1.5.0
gitdb==4.0.5
GitPython==3.1.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.30.0
h11==0.8.1
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.6.23
httplib2==0.18.1
httptools==0.1.1
httpx==0.9.3
humanfriendly==8.2
hyperframe==5.2.0
idna==2.9
idna-ssl==1.1.0
immutables==0.14
importlib-metadata==1.6.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsonpickle==1.4.1
jsonschema==3.2.0
kafka-python==1.4.7
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mattermostwrapper==2.2
multidict==4.7.6
mysql-connector==2.2.9
mysqlclient==2.0.1
networkx==2.4
numpy==1.19.0
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.5
pika==1.1.0
pkg-resources==0.0.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
PyJWT==1.7.1
pykwalify==1.7.0
pymongo==3.8.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.12.1
python-socketio==4.5.1
python-telegram-bot==12.8
pytz==2020.1
PyYAML==5.3.1
questionary==1.5.2
rasa==1.10.3
rasa-sdk==1.10.2
rasa-x==0.29.3
ratelim==0.1.6
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rocketchat-API==1.3.1
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
sanic==19.12.2
Sanic-Cors==0.10.0.post3
sanic-jwt==1.3.2
Sanic-Plugins-Framework==0.9.2
scikit-learn==0.22.2.post1
scipy==1.5.0
six==1.15.0
sklearn-crfsuite==0.3.6
slackclient==2.7.2
smmap==3.0.4
sniffio==1.1.0
SQLAlchemy==1.3.18
tabulate==0.8.7
tensorboard==2.1.1
tensorflow==2.1.1
tensorflow-addons==0.7.1
tensorflow-estimator==2.1.0
tensorflow-hub==0.8.0
tensorflow-probability==0.9.0
termcolor==1.1.0
terminaltables==3.1.0
tornado==6.0.4
tqdm==4.45.0
twilio==6.26.3
typing-extensions==3.7.4.2
tzlocal==2.1
ujson==1.35
urllib3==1.25.9
uvloop==0.14.0
wcwidth==0.2.5
webexteamssdk==1.3
websockets==8.1
Werkzeug==0.16.1
wrapt==1.12.1
yarl==1.4.2
zipp==3.1.0
